Refactor config setup (env, database, session, test, jwt)
