+++ app/tests/conftest.py
@@ -1,9 +1,7 @@
 import pytest
 from flask import g
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker, Session
-from app.app import create_app
-from app.model.database import Base
+from sqlalchemy.orm import Session
+from app.model.database import Base, engine, SessionLocal
 from app.model.models import User, Product, Order, OrderItem
 from app.core.auth_utils import generate_token
 from werkzeug.security import generate_password_hash
@@ -12,14 +10,12 @@ from app.services.order_utils import create_new_order, get_orderitems_all
 
 from flask.testing import FlaskClient
 from typing import Tuple, Generator, List, Dict, Any
+from app import create_app
+import os
 
+# Force TestConfig à la creation de app
+os.environ["FLASK_ENV"] = "testing"
 
-# '''TEMPORAIRE: engine mémoire crée avant import de create_app'''
-# engine = create_engine("sqlite:///:memory:", echo=False)
-# SessionLocal = sessionmaker(bind=engine)
-# Base.metadata.create_all(bind=engine)
-# from app.app import create_app
-# from flask import g
 
 @pytest.fixture(scope="session")
 def setup_db() -> Generator[Tuple, None, None]:
@@ -29,16 +25,11 @@ def setup_db() -> Generator[Tuple, None, None]:
 
     Retourne un tuple :
         - Flask app (configurée pour les tests)
-        - sessionmaker (lié à la base en mémoire)
+        - SessionLocal (définie fois dans database.py)
     """
     app = create_app()
     app.config["TESTING"] = True
-
-    ''' MUET TEMPORAIRE LE TEMPS DE MIGRER ET INTEGRER MULTI-ENV'''
-    engine = create_engine("sqlite:///:memory:", echo=False)
-
-    # -> autoflush=True, autocommit=True pour souci commit() en PROD
-    SessionLocal = sessionmaker(bind=engine)
+   
     Base.metadata.create_all(bind=engine)
     yield app, SessionLocal
 
@@ -105,14 +96,6 @@ def client_token(test_client) -> str:
     session.add(user)
     session.commit()
     return generate_token(user)
-
-
-# @pytest.fixture(scope="function")
-# def visitor_only(test_client):
-#     """
-#     Crée un utilisateur sans authentification ni autorisation.
-#     """
-#     return None
 
 
 @pytest.fixture(scope="function")
@@ -189,3 +172,11 @@ def feed_order(test_client, client_token, feed_product) -> Dict[str, Any]:
         "utilisateur_id": user.id,
         "email": user.email
         }
+
+
+# @pytest.fixture(scope="function")
+# def visitor_only(test_client):
+#     """
+#     Crée un utilisateur sans authentification ni autorisation.
+#     """
+#     return None
